[Minigame] SampleLetterProvider fix

+++ Assets/_games/_common/_scripts/_mustbeimplementedbyhub/SampleLetterProvider.cs
@@ -7,7 +7,7 @@ namespace EA4S
     /// </summary>
     public class SampleLetterProvider : ILivingLetterDataProvider
     {
-        public SampleLetterProvider(int difficulty)
+        public SampleLetterProvider(float difficulty)
         {
 
         }
